Odered Tree Minors
git-svn-id: https://scm.gforge.inria.fr/svn/rtcelog/software/JBenchmarkerCE@1018 47090b3f-bbb5-42d7-8af4-003170c53eba

+++ src/main/java/jbenchmarker/logoot/BoundaryStrategy.java
@@ -33,7 +33,7 @@ public class BoundaryStrategy extends LogootStrategy {
     private final long bound;
     private final BigInteger boundBI;
 
-    public BoundaryStrategy(int bound) {
+    public BoundaryStrategy(long bound) {
         this.bound = bound;
         this.boundBI = BigInteger.valueOf(bound);
     }
